# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Car Reservation API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/dokters": {
            "get": {
                "tags": [
                    "Dokter"
                ],
                "summary": "Get all doctors",
                "operationId": "db70b3686ffa729caed6aec1d702db53",
                "responses": {
                    "200": {
                        "description": "Dokter retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Dokter"
                ],
                "summary": "Create a new doctor",
                "operationId": "2ee20325e288366d4de6e926c9de9cd1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "nama",
                                    "spesialis",
                                    "no_telp",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "spesialis": {
                                        "type": "string"
                                    },
                                    "no_telp": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dokter created successfully"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dokters/{id}": {
            "get": {
                "tags": [
                    "Dokter"
                ],
                "summary": "Get doctor by ID",
                "operationId": "7a2be6e52059331444507c133ba43152",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dokter retrieved successfully"
                    },
                    "404": {
                        "description": "Dokter not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Dokter"
                ],
                "summary": "Update a doctor by ID",
                "operationId": "b4b4eb4191f7c4c1d1a8a5c8d821c079",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "nama",
                                    "spesialis",
                                    "no_telp",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "spesialis": {
                                        "type": "string"
                                    },
                                    "no_telp": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dokter updated successfully"
                    },
                    "404": {
                        "description": "Dokter not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Dokter"
                ],
                "summary": "Delete doctor by ID",
                "operationId": "48de73bb6dd2a357941c9ff3296754f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dokter deleted successfully"
                    },
                    "404": {
                        "description": "Dokter not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dokters/spesialis/{spesialis}": {
            "get": {
                "tags": [
                    "Dokter"
                ],
                "summary": "Ambil daftar dokter berdasarkan spesialis",
                "operationId": "0a1bf1e1a009ca881f4d992e70d51a8b",
                "parameters": [
                    {
                        "name": "spesialis",
                        "in": "path",
                        "description": "Spesialis dokter, misalnya: anak, gigi, umum",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses mengambil data dokter",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data dokter ditemukan"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "dr. Siti"
                                                    },
                                                    "spesialis": {
                                                        "type": "string",
                                                        "example": "gigi"
                                                    },
                                                    "no_telp": {
                                                        "type": "string",
                                                        "example": "081234567890"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "dr.siti@example.com"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dokter tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Masukkan token dalam format: Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}